# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is an Official Documentation for Go REST API with Fiber Framework",
        "title": "Go REST API with Fiber Framework",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Fatih Rizqon",
            "email": "fatihrizqon@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "127.0.0.1:3000",
    "basePath": "/",
    "paths": {
...
        "/api/v1/companies": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve all company records with pagination",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Companies"
                ],
                "summary": "Get all companies",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Search keyword",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page size",
                        "name": "page_size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Sort column (name, legal_name, email, status, created_at, updated_at)",
                        "name": "sort",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Sort direction (asc, desc)",
                        "name": "order",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by status",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by email verified (true, false)",
                        "name": "verified",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved all records.",
                        "schema": {
                            "$ref": "#/definitions/response.JSON"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.JSON"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Store a new company record",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Companies"
                ],
                "summary": "Create company",
                "parameters": [
                    {
                        "description": "Company Create Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.CompanyCreateRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "A new record has been stored.",
                        "schema": {
                            "$ref": "#/definitions/response.JSON"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/response.JSON"
                        }
                    }
                }
            }
        },
        "/api/v1/companies/active": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns the company that is currently active in this session",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Companies"
                ],
                "summary": "Get current active company",
                "responses": {
                    "200": {
                        "description": "Successfully retrieved active company.",
                        "schema": {
                            "$ref": "#/definitions/response.JSON"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.JSON"
                        }
                    },
                    "404": {
                        "description": "No active company selected",
                        "schema": {
                            "$ref": "#/definitions/response.JSON"
                        }
                    }
                }
            }
        },
        "/api/v1/companies/mine": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "Company"
                ],
                "summary": "Get all companies the authenticated user belongs to",
                "responses": {}
            }
        },
        "/api/v1/companies/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve a single company by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Companies"
                ],
                "summary": "Get company by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Company ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved selected record.",
                        "schema": {
                            "$ref": "#/definitions/response.JSON"
                        }
                    },
                    "404": {
                        "description": "Company not found",
                        "schema": {
                            "$ref": "#/definitions/response.JSON"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update company data by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Companies"
                ],
                "summary": "Update company",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Company ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Company Update Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.CompanyUpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Selected record has been updated.",
                        "schema": {
                            "$ref": "#/definitions/response.JSON"
                        }
                    },
                    "404": {
                        "description": "Company not found",
                        "schema": {
                            "$ref": "#/definitions/response.JSON"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Remove a company record by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Companies"
                ],
                "summary": "Delete company",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Company ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Selected record has been deleted.",
                        "schema": {
                            "$ref": "#/definitions/response.JSON"
                        }
                    },
                    "404": {
                        "description": "Company not found",
                        "schema": {
                            "$ref": "#/definitions/response.JSON"
                        }
                    }
                }
            }
        },
        "/api/v1/companies/{id}/members": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "Company"
                ],
                "summary": "Assign member to company",
                "responses": {}
            }
        },
        "/api/v1/companies/{id}/members/{userId}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "Company"
                ],
                "summary": "Remove member from company",
                "responses": {}
            }
        },
        "/api/v1/companies/{id}/members/{userId}/role": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "Company"
                ],
                "summary": "Update member role",
                "responses": {}
            }
        },
        "/api/v1/companies/{id}/select": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Set the active company for the current session. Call this again to switch company.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Companies"
                ],
                "summary": "Select or switch active company",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Company ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Active company has been set.",
                        "schema": {
                            "$ref": "#/definitions/response.JSON"
                        }
                    },
                    "400": {
                        "description": "Invalid company ID",
                        "schema": {
                            "$ref": "#/definitions/response.JSON"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.JSON"
                        }
                    },
                    "403": {
                        "description": "Not a member of this company",
                        "schema": {
                            "$ref": "#/definitions/response.JSON"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "request.CompanyCreateRequest": {
            "type": "object",
            "required": [
                "email",
                "name",
                "phone"
            ],
            "properties": {
                "address": {
                    "type": "string"
                },
                "currency": {
                    "type": "string",
                    "maxLength": 10
                },
                "email": {
                    "type": "string",
                    "maxLength": 150,
                    "minLength": 1
                },
                "industry": {
                    "type": "string",
                    "maxLength": 100
                },
                "legal_name": {
                    "type": "string",
                    "maxLength": 200
                },
                "name": {
                    "type": "string",
                    "maxLength": 150,
                    "minLength": 1
                },
                "phone": {
                    "type": "string",
                    "maxLength": 30,
                    "minLength": 1
                },
                "tax_id": {
                    "type": "string",
                    "maxLength": 50
                }
            }
        },
        "request.CompanyUpdateRequest": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "address": {
                    "type": "string"
                },
                "currency": {
                    "type": "string",
                    "maxLength": 10
                },
                "email": {
                    "type": "string",
                    "maxLength": 150
                },
                "id": {
                    "type": "string"
                },
                "industry": {
                    "type": "string",
                    "maxLength": 100
                },
                "legal_name": {
                    "type": "string",
                    "maxLength": 200
                },
                "name": {
                    "type": "string",
                    "maxLength": 150,
                    "minLength": 1
                },
                "phone": {
                    "type": "string",
                    "maxLength": 30
                },
                "tax_id": {
                    "type": "string",
                    "maxLength": 50
                }
            }
        },
...
        "response.JSON": {
            "type": "object",
            "properties": {
                "data": {},
                "errors": {},
                "message": {
                    "type": "string"
                },
                "meta": {
                    "$ref": "#/definitions/response.Meta"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
        "response.Links": {
            "type": "object",
            "properties": {
                "current_page": {
                    "type": "string"
                },
                "first_page": {
                    "type": "string"
                },
                "last_page": {
                    "type": "string"
                },
                "next_page": {
                    "type": "string"
                },
                "prev_page": {
                    "type": "string"
                }
            }
        },
        "response.Meta": {
            "type": "object",
            "properties": {
                "info": {
                    "type": "string"
                },
                "links": {
                    "$ref": "#/definitions/response.Links"
                },
                "page": {
                    "type": "integer"
                },
                "page_size": {
                    "type": "integer"
                },
                "search": {
                    "type": "string"
                },
                "total_count": {
                    "type": "integer"
                },
                "total_pages": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}